Fully integrate patch from Sony concerning the plain old data findNPoints
The goal is to be able to call findNPoints() without having a std::vector around.
Whitespace changes removed.
Made both use the same query structure in KdTree.h (now the stl query is in terms of a POD query)
Add a test for both calls in the testkdtree test.

Also the data type check on dataMultiple now is disabled if you send in a void* pointer.