Bugfix: PTC file normals sometimes encode as null
Normals in PTC files are compressed into two unsigned short values,
the z coordinate and the angle phi in the xy plane.  Encoded normals are
unit length, except for the special value of z == phi == 65535 which
is used for zero length normals.  This special value can arise naturally
during compression of non-null normals close to (0,0,1) and must be
excluded explicitly by changing phi to something else.  Anything else
will do since phi is redundent when z is a maximum, so phi = 0 is
chosen.