doc: fix doxygen build
HTML_TARGET is undefined.  Name the target "doc".

Ensure that we depend on Doxyfile in the binary directory, not the
source directory, as it is a generated file.

Use add_custom_command() to avoid needing to re-run doxygen each time.
We create a "doxygen.txt" file which is used as a dependency stub file.
The presence of this file prevents rebuilds.