io/MC.cpp: fix comparison against '\0'
charArray is a char** so use `nullptr` when comparing against it.
Similarly, this funciton only cares to check how many char* buffers have
been allocated inside the char**, so the conditional should check
whether any of the members are nullptr.

We do not check for '\0' inside the string because empty strings are
still valid strings.

Related-to: https://www.viva64.com/en/b/0455/