Toward FXSOFT-1129: Add Partio::merge() function
void merge(ParticlesDataMutable& base,
           const ParticlesData& delta,
           const std::string& identifier=std::string());

Given a ParticleSetMutable, merges it with a second ParticleSet, copying
particles and attributes that align with the base particle set. If an
identifier is provided, that will be used as a key to replace the
particle in the base set with the particle in the second set with the
same identifier attribute value. If the identifier is not provided or
the particle's attribute value is not found in the base set, a new
particle is added. If used, the identifier must be a single INT.