Fixes FXSOFT-1185: Spreadsheet button not working correctly
Refactored the partinspect a little bit to encapsulate data into the
dialog itself rather than as global values within the module.

Public functions still exist to utilize a single copy of the spreadsheet
(really, a single copy each of particle vs attr data). But this is no
longer a necessity. However due to the reuse of the spreadsheet between
xgen and aurora GUI's, the public functions remained a better solve for
things like live frame changes.