Toward FXSOFT-1382: getting crash in the partio api
The fundamental cause of this bug was using the particle ID as the
particle index. In the specific case, an ID was 17, but there were only
9 particles in the file. Setting a particle attribute at an index
greater than the number of particles in the file causes a memory write
to an unrelated area, essentially inflicting mass damage and crashes.

Added a couple of defensive checks to help with this.

Most importantly, on the call to retrieve the write location, the
particleIndex is now validated against the number of particles before
returning a pointer for writing. If beyond the end of the range, a
nullptr is returned.

Second, this had to be handled in the consuming locations, including the
set() function and the swig interface.

Finally, if the provided file is an empty file, I've added a check on
the input stream that the read itself actually succeeded.