Partio::clone: Add optional map arg for attribute renaming
Add an optional map<string, string> ptr argument to clone and
cloneSchema, which defaults to nullptr. If provided, the map will be
used to remap attribute names during the clone functions.

This was implemented in an effort towards deprecating bgeos in the
aurora workflow and moving towards UsdGeomPoints-based particle sources.
Bgeo and usdGeomPoints have different attribute naming conventions which
must be resolved so that particle sources of both types can be merged.
We are using the UsdGeomPoints conventions, so bgeo data must have it's
attribute names converted, which requires cloning the data completely
and changing attribute names in the process.