support merging particle sets based on 2 different attributes
Previously, you could merge particle sets and override a particle
when the value of a specified attribute (typ. 'ids') was identical.  Because
ids can sometimes be unintentionally duplicated, an additional attribute
can be specified when performing the matching.  When used, a particle is
overridden only when the values of both attributes are identical.

toward: LOOK-1965 Duplicate particle IDs in baked aurora xgenAuroraFiles