partinfo crash fix
In the example where I discovered this crash, there were indexed strings
for "caf" but no values. For a particle that had a "caf" attribute, it
has to have a value so the value was 0. But indexing into the
non-existent indexed-string array was, as expected, causing a crash.
Just needed to add a simple bounds check.